==> 04-hyperparameter-tuning-experiment.ipynb <==
# Titanic challenge with XGBoost - Hyperparams tuning - SM Experiments

This notebook resolves Titanic challenge using XGBoost model. The target is to get familiar with AWS Sagemaker and one of the most popular model - eXtreme Gradient Boosting.
The original challenge is defined at https://www.kaggle.com/c/titanic/data

This notebook experiments Hyperparams tuning with AWS SageMaker, analyze the results, and then attach created traning jobs to SageMaker Experiment.

Note: At the writing time, some Experiment features such as Create Trial Component can only be invoked from within an Amazon SageMaker managed environment. This includes Amazon SageMaker training jobs, processing jobs, transform jobs, and Amazon SageMaker notebooks. 

This notebook is run in a SageMaker Notebook.

## 1. Preparation

# import libraries
import boto3
import sagemaker

from sagemaker.tuner import IntegerParameter, CategoricalParameter, ContinuousParameter, HyperparameterTuner

from sagemaker import HyperparameterTuningJobAnalytics, Session

from smexperiments.experiment import Experiment
from smexperiments.search_expression import Filter, Operator, SearchExpression
from smexperiments.trial import Trial
from smexperiments.trial_component import TrialComponent
from smexperiments.tracker import Tracker

from sagemaker import get_execution_role

import time

# Define bucket name and prefix
bucket = '<bucket-name>' 
prefix = 'prefix'

# Define IAM role and sagemaker client
boto_session = boto3.Session()
session = sagemaker.Session(boto_session=boto_session)

# Get the current role of the SageMaker Notebook
role = get_execution_role(session)

# define the local data path
train_data_file = './data/processed/exp-raw/train.csv'
validation_data_file = './data/processed/exp-raw/validation.csv'
test_data_file = './data/processed/exp-raw/test.csv'

# upload local data into s3
s3_train_uri = session.upload_data(path=train_data_file, bucket=bucket, key_prefix='/'.join((prefix, 'basic/train')))
s3_validate_uri = session.upload_data(path=validation_data_file, bucket=bucket, key_prefix='/'.join((prefix, 'basic/validation')))
s3_test_uri = session.upload_data(path=test_data_file, bucket=bucket, key_prefix='/'.join((prefix, 'basic/test')))


with Tracker.create(display_name="Basic-Preprocessing", boto3_session=boto_session) as basic_tracker:
    basic_tracker.log_parameters({
        "sex": 'dummies',
        "pclass": 'dummies',
        "name": 'drop',
        "ticket": 'drop',
        "cabin": 'drop',
        "embarked": 'dummies'
    })
    # we can log the s3 uri to the dataset we just uploaded
    basic_tracker.log_input(name="titanic-training-dataset", media_type="s3/uri", value=s3_train_uri)
    basic_tracker.log_input(name="titanic-validate-dataset", media_type="s3/uri", value=s3_validate_uri)

## 2. Hyperparam tuning

#### XGBoosting Hyperparameters ####
#### General params
- eval_metric: Evaluation metrics for validation data. A default metric is assigned according to the objective 
- objective: Specifies the learning task and the corresponding learning objective.
- num_round: The number of rounds to run the training = number of estimators.
- rate_drop: The dropout rate that specifies the fraction of previous trees to drop during the dropout.
- tweedie_variance_power: Parameter that controls the variance of the Tweedie distribution.

#### Tree booster params
- eta: Step size shrinkage used in updates to prevent overfitting. After each boosting step, you can directly get the weights of new features. The eta parameter actually shrinks the feature weights to make the boosting process more conservative.
- min_child_weight: Minimum sum of instance weight (hessian) needed in a child. If the tree partition step results in a leaf node with the sum of instance weight less than min_child_weight, the building process gives up further partitioning. In linear regression models, this simply corresponds to a minimum number of instances needed in each node. The larger the algorithm, the more conservative it is.
- alpha: L1 regularization term on weights. Increasing this value makes models more conservative.
- max_depth: Maximum depth of a tree. Increasing this value makes the model more complex and likely to be overfit. 0 indicates no limit. 

# Get the XGBoost image uri corresponding to the current region
from sagemaker import image_uris
container = image_uris.retrieve('xgboost', session.boto_region_name, 'latest')

# Define the estimator and hyperparams
xgb = sagemaker.estimator.Estimator(container,
                                    role, # role to be attached to instance to access data
                                    instance_count=1, 
                                    instance_type='ml.m4.xlarge',
                                    output_path='s3://{}/{}/output'.format(bucket, prefix),
                                    use_spot_instances=True,
                                    max_run=400,
                                    max_wait=600)

xgb.set_hyperparameters(eval_metric='auc',
                        objective='binary:logistic',
                        early_stopping_round=10)

# Define hyperparams for tuning
hyperparameter_ranges = {'eta': ContinuousParameter(0, 1),
                        'min_child_weight': ContinuousParameter(1, 10),
                        'alpha': ContinuousParameter(0, 2),
                        'max_depth': IntegerParameter(1, 10),
                        'num_round':IntegerParameter(50, 500)}

objective_metric_name = 'validation:auc'

tuner = HyperparameterTuner(xgb,
                            objective_metric_name,
                            hyperparameter_ranges,
                            max_jobs=12,
                            max_parallel_jobs=4)

# Define the input points for xgboost
s3_input_train = sagemaker.inputs.TrainingInput(s3_data=s3_train_uri, content_type='csv')
s3_input_validation = sagemaker.inputs.TrainingInput(s3_data=s3_validate_uri, content_type='csv')

tuner.fit({'train': s3_input_train, 'validation': s3_input_validation}, include_cls_metadata=False)

## 3. Evaluation

import pandas as pd

tuning_job_name = 'xgboost-201216-1247'

# check the status of tuning job
tuning_job = boto3.client('sagemaker').describe_hyper_parameter_tuning_job(HyperParameterTuningJobName=tuning_job_name)

status = tuning_job['HyperParameterTuningJobStatus']
if status != 'Completed':
    print('Reminder: the tuning job has not been completed.')
    
job_count = tuning_job['TrainingJobStatusCounters']['Completed']
print("%d training jobs have completed" % job_count)

is_minimize = (tuning_job['HyperParameterTuningJobConfig']['HyperParameterTuningJobObjective']['Type'] != 'Maximize')
objective_name = tuning_job['HyperParameterTuningJobConfig']['HyperParameterTuningJobObjective']['MetricName']

tuner_result = sagemaker.HyperparameterTuningJobAnalytics(tuning_job_name)

full_df = tuner_result.dataframe()

if len(full_df) > 0:
    df = full_df[full_df['FinalObjectiveValue'] > -float('inf')]
    if len(df) > 0:
        df = df.sort_values('FinalObjectiveValue', ascending=is_minimize)
        print("Number of training jobs with valid objective: %d" % len(df))
        print({"lowest":min(df['FinalObjectiveValue']),"highest": max(df['FinalObjectiveValue'])})
        pd.set_option('display.max_colwidth', -1)  # Don't truncate TrainingJobName        
    else:
        print("No training jobs have reported valid results yet.")
        
df

import bokeh
import bokeh.io
bokeh.io.output_notebook()
from bokeh.plotting import figure, show
from bokeh.models import HoverTool

class HoverHelper():

    def __init__(self, tuning_analytics):
        self.tuner = tuning_analytics

    def hovertool(self):
        tooltips = [
            ("FinalObjectiveValue", "@FinalObjectiveValue"),
            ("TrainingJobName", "@TrainingJobName"),
        ]
        for k in self.tuner.tuning_ranges.keys():
            tooltips.append( (k, "@{%s}" % k) )

        ht = HoverTool(tooltips=tooltips)
        return ht

    def tools(self, standard_tools='pan,crosshair,wheel_zoom,zoom_in,zoom_out,undo,reset'):
        return [self.hovertool(), standard_tools]

hover = HoverHelper(tuner_result)

p = figure(plot_width=900, plot_height=400, tools=hover.tools(), x_axis_type='datetime')
p.circle(source=df, x='TrainingStartTime', y='FinalObjectiveValue')
show(p)

ranges = tuner_result.tuning_ranges
figures = []
for hp_name, hp_range in ranges.items():
    categorical_args = {}
    if hp_range.get('Values'):
        # This is marked as categorical.  Check if all options are actually numbers.
        def is_num(x):
            try:
                float(x)
                return 1
            except:
                return 0           
        vals = hp_range['Values']
        if sum([is_num(x) for x in vals]) == len(vals):
            # Bokeh has issues plotting a "categorical" range that's actually numeric, so plot as numeric
            print("Hyperparameter %s is tuned as categorical, but all values are numeric" % hp_name)
        else:
            # Set up extra options for plotting categoricals.  A bit tricky when they're actually numbers.
            categorical_args['x_range'] = vals

    # Now plot it
    p = figure(plot_width=500, plot_height=500, 
               title="Objective vs %s" % hp_name,
               tools=hover.tools(),
               x_axis_label=hp_name, y_axis_label=objective_name,
               **categorical_args)
    p.circle(source=df, x=hp_name, y='FinalObjectiveValue')
    figures.append(p)
show(bokeh.layouts.Column(*figures))

# tuner.best_training_job()
tuner.best_estimator().hyperparameters()

### 4. Attach tuning job to experiment again

experiment_name = "titanic-experiment"
trial_name = tuning_job_name + "-tuner"

# Create/load an experiment to manage all attempts in titanic problem
try:
    titanic_experiment = Experiment.load(experiment_name=experiment_name)
except Exception as ex:
    print("Create a new experiment")
    if "ResourceNotFound" in str(ex):
        titanic_experiment = Experiment.create(
            experiment_name = experiment_name, 
            description = "Classification of survivors titanic")
            
titanic_experiment

# Define the trial using basic dataset
basic_trial = Trial.create(
    trial_name = 'basic-' + trial_name, 
    experiment_name = titanic_experiment.experiment_name
)

basic_trial.add_trial_component(basic_tracker.trial_component)

tuning_job = boto3.client('sagemaker').describe_hyper_parameter_tuning_job(HyperParameterTuningJobName=tuning_job_name)
# tuning_job

We can add all training jobs to the trial, each training job is a trial component. [1] 
Or we can add only the best traning job to the trial [2].
The following code shows how to add all traning jobs of this hyperparam tuning job into a trial ([1]).

# get the trial components derived from the training jobs
from datetime import datetime, timezone

creation_time = tuning_job["CreationTime"]
creation_time = creation_time.astimezone(timezone.utc)
creation_time = creation_time.strftime("%Y-%m-%dT%H:%M:%SZ")

created_after_filter = Filter(
    name="CreationTime",
    operator=Operator.GREATER_THAN_OR_EQUAL,
    value=str(creation_time),
)

# the training job names contain the tuning job name (and the training job name is in the source arn)
source_arn_filter = Filter(
    name="Source.SourceArn", operator=Operator.CONTAINS, value=tuning_job_name
)
source_type_filter = Filter(
    name="Source.SourceType", operator=Operator.EQUALS, value="SageMakerTrainingJob"
)

search_expression = SearchExpression(
    filters=[created_after_filter, source_arn_filter, source_type_filter]
)

# search iterates over every page of results by default
trial_component_search_results = list(
    TrialComponent.search(search_expression=search_expression)
)
print(f"Found {len(trial_component_search_results)} trial components.")
# type(trial_component_search_results[0])
# trial_component_search_results[0]

# associate the trial components with the trial
for tc in trial_component_search_results:
    print(
        f"Associating trial component {tc.trial_component_name} with trial {basic_trial.trial_name}."
    )
    basic_trial.add_trial_component(tc.trial_component_name)
    # sleep to avoid throttling
    time.sleep(0.5)

titanic_experiment.list_trials()

for trial_summary in titanic_experiment.list_trials():
        trial = Trial.load(trial_name=trial_summary.trial_name)
        print(trial_summary)

def cleanup_sme_sdk(experiment):
    for trial_summary in experiment.list_trials():
        trial = Trial.load(trial_name=trial_summary.trial_name)
        for trial_component_summary in trial.list_trial_components():
            tc = TrialComponent.load(
                trial_component_name=trial_component_summary.trial_component_name)
            trial.remove_trial_component(tc)
            try:
                # comment out to keep trial components
                tc.delete()
            except:
                # tc is associated with another trial
                continue
            # to prevent throttling
            time.sleep(.5)
        trial.delete()
        experiment_name = experiment.experiment_name
    experiment.delete()
    print(f"\nExperiment {experiment_name} deleted")
